+++ src/hotdogstand.js
@@ -1,5 +1,5 @@
 /*
- * Hot Dog Stand v0.1.0
+ * Hot Dog Stand v0.2.0
  * Copyright 2015, Steve Richey
  * https://github.com/steverichey/hot-dog-stand
  * Free to use under the MIT license.
@@ -13,20 +13,23 @@
 	var hot = {
 		dog: {
 			get stand() {
-				var everything = document.querySelector("*");
-				everything.style.setProperty ("background", "red", "important");
-				everything.style.setProperty ("color", "yellow", "important");
+				var el = document.querySelectorAll("*"), i;
+                
+                for (i = 0; i < el.length; i++) {
+                    el[i].style.setProperty ("background", "red", "important");
+                    el[i].style.setProperty ("color", "yellow", "important");
+                }
 			}
 		}
 	}
 
 	// i stole this from PIXI
-	if (typeof exports !== 'undefined') {
-        if (typeof module !== 'undefined' && module.exports) {
+	if (typeof exports !== "undefined") {
+        if (typeof module !== "undefined" && module.exports) {
             exports = module.exports = hot;
         }
         exports.hot = hot;
-    } else if (typeof define !== 'undefined' && define.amd) {
+    } else if (typeof define !== "undefined" && define.amd) {
         define(hot);
     } else {
         root.hot = hot;