+++ changelog.md
@@ -1,3 +1,8 @@
+# v0.2.0
+
+* Improved JavaScript framework
+* Chrome Extension
+
 # v0.1.0
 
 * Travis, Circle, AppVeyor, Semaphore, CodeShip testing